日常

+++ UIEditor/PackImage.xaml.cs
@@ -125,32 +125,43 @@ namespace UIEditor
 				System.Drawing.Graphics g = System.Drawing.Graphics.FromImage(m_tgaImg);
 				g.Clear(System.Drawing.Color.FromArgb(0x00, 0x00, 0x00, 0x00));
 
-				string tmpPath = System.IO.Path.GetFileNameWithoutExtension(m_parent.m_openedFile.m_path);
+				string pngPath = m_parent.m_openedFile.m_path.Remove(m_parent.m_openedFile.m_path.LastIndexOf("."));
 
-				if (tmpPath.LastIndexOf("_wpf") >= 0)
-				{
-					tmpPath = tmpPath.Remove(tmpPath.LastIndexOf("_wpf"));
-				}
+// 				if (tmpPath.LastIndexOf("_wpf") >= 0)
+// 				{
+// 					tmpPath = tmpPath.Remove(tmpPath.LastIndexOf("_wpf"));
+// 				}
 				foreach (KeyValuePair<string, System.Drawing.Rectangle> pairImgRect in m_mapImgRect)
 				{
 					addPicToGraphics(
-						MainWindow.s_pW.m_projPath + "\\images\\" + tmpPath + 
-							"\\" + pairImgRect.Key + ".png",
+						pngPath + "\\" + pairImgRect.Key + ".png",
 						pairImgRect.Value,
 						g);
 				}
 				g.Dispose();
-				DevIL.DevIL.SaveBitmap("E:\\tmp\\" + tmpPath + ".tga", m_tgaImg);
-				ip = m_tgaImg.GetHbitmap();
-				m_imgSource = System.Windows.Interop.Imaging.CreateBitmapSourceFromHBitmap(
-					ip, IntPtr.Zero, Int32Rect.Empty,
-					System.Windows.Media.Imaging.BitmapSizeOptions.FromEmptyOptions());
-				MainWindow.DeleteObject(ip);
+				string tgaPath = pngPath + ".tga";
+				if(System.IO.File.Exists(tgaPath))
+				{
+					System.IO.File.Delete(tgaPath);
+				}
+				DevIL.DevIL.SaveBitmap(tgaPath, m_tgaImg);
+				if(m_imgHeight <= 4096)
+				{
+					ip = m_tgaImg.GetHbitmap();
+					m_imgSource = System.Windows.Interop.Imaging.CreateBitmapSourceFromHBitmap(
+						ip, IntPtr.Zero, Int32Rect.Empty,
+						System.Windows.Media.Imaging.BitmapSizeOptions.FromEmptyOptions());
+					MainWindow.DeleteObject(ip);
 
-				System.Windows.Controls.Image cImg = new System.Windows.Controls.Image();
-				cImg.Source = m_imgSource;
-				cImg.Stretch = Stretch.Uniform;
-				mx_canvas.Children.Insert(0, cImg);
+					System.Windows.Controls.Image cImg = new System.Windows.Controls.Image();
+					cImg.Source = m_imgSource;
+					cImg.Stretch = Stretch.Uniform;
+					mx_canvas.Children.Insert(0, cImg);
+				}
+				else
+				{
+					MainWindow.s_pW.mx_debug.Text += "<警告>图片尺寸过大，不提供预览功能\r\n";
+				}
 				m_loaded = true;
 			}
 		}
@@ -189,12 +200,14 @@ namespace UIEditor
 				if (pairImgRect.Value.Contains(x, y))
 				{
 					mx_overPath.Visibility = System.Windows.Visibility.Visible;
-					mx_overPath.Data = new RectangleGeometry(new Rect(
-						pairImgRect.Value.X,
-						pairImgRect.Value.Y,
-						pairImgRect.Value.Width,
-						pairImgRect.Value.Height
-					));
+					mx_overPath.Data = new RectangleGeometry(
+						new Rect(
+							pairImgRect.Value.X,
+							pairImgRect.Value.Y,
+							pairImgRect.Value.Width,
+							pairImgRect.Value.Height
+						)
+					);
 
 					return;
 				}
